#40 - bufixes2

+++ api/DiscordBotConnection.ts
@@ -34,15 +34,17 @@ const DiscordBotConnection = async () => {
 
     const discordleChannelId = channels.find(
       (c) => c.name === "daily-discordle" && c.type === ChannelType.GuildText
-    ).id;
+    )?.id;
 
-    await request(
-      `https://discord.com/api/v10/channels/${discordleChannelId}`,
-      {
-        method: "DELETE",
-        headers: { authorization: `Bot ${process.env.BOT_TOKEN}` },
-      }
-    );
+    if (discordleChannelId) {
+      await request(
+        `https://discord.com/api/v10/channels/${discordleChannelId}`,
+        {
+          method: "DELETE",
+          headers: { authorization: `Bot ${process.env.BOT_TOKEN}` },
+        }
+      );
+    }
 
     const filteredChannels = await Promise.all(
       channels.map(async (channel) => {
@@ -101,7 +103,7 @@ const DiscordBotConnection = async () => {
 
     await guild.channels
       .create({
-        name: "daily-discordle",
+        name: "Daily Discordle",
         type: ChannelType.GuildText,
         permissionOverwrites: [
           {
